requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
Werkzeug==3.0.6
Jinja2==3.1.6
python-dateutil==2.8.2
Flask-Moment==1.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.10
